renaming http-api -> http, gateway spec tests
License: MIT
