+++ package.json
@@ -27,6 +27,7 @@
     "test:unit:node": "gulp test:node",
     "test:unit:node:core": "TEST=core npm run test:unit:node",
     "test:unit:node:http": "TEST=http npm run test:unit:node",
+    "test:unit:node:gateway": "TEST=gateway npm run test:unit:node",
     "test:unit:node:cli": "TEST=cli npm run test:unit:node",
     "test:unit:browser": "gulp test:browser",
     "test:interop": "npm run test:interop:node",
