+++ src/http/api/resources/bootstrap.js
