+++ src/http/api/routes/files.js
@@ -4,7 +4,6 @@ const resources = require('./../resources')
 
 module.exports = (server) => {
   const api = server.select('API')
-  const gateway = server.select('Gateway')
 
   api.route({
     // TODO fix method
@@ -40,17 +39,6 @@ module.exports = (server) => {
         output: 'stream'
       },
       handler: resources.files.add.handler
-    }
-  })
-
-  gateway.route({
-    method: '*',
-    path: '/ipfs/{hash*}',
-    config: {
-      pre: [
-        { method: resources.files.gateway.checkHash, assign: 'args' }
-      ],
-      handler: resources.files.gateway.handler
     }
   })
 }
