+++ src/http/gateway/resources/index.js
@@ -1,0 +1,3 @@
+'use strict'
+
+exports.gateway = require('./gateway')
