+++ src/http/gateway/routes/index.js
@@ -1,0 +1,5 @@
+'use strict'
+
+module.exports = (server) => {
+  require('./gateway')(server)
+}
