+++ src/http/gateway/utils/path.js
