+++ src/http/gateway/utils/style.js
