+++ src/http/index.js
@@ -106,7 +106,9 @@ function HttpApi (repo, config, cliArgs) {
           errorHandler(this, this.server)
 
           // load routes
-          require('./routes')(this.server)
+          require('./api/routes')(this.server)
+          // load gateway routes
+          require('./gateway/routes')(this.server)
 
           // Set default headers
           setHeader(this.server,
