+++ test/gateway/index.js
@@ -7,12 +7,12 @@ const dirtyChai = require('dirty-chai')
 const expect = chai.expect
 chai.use(dirtyChai)
 const API = require('../../src/http')
-const APIctl = require('ipfs-api')
+// const APIctl = require('ipfs-api')
 const ncp = require('ncp').ncp
 const path = require('path')
 const clean = require('../utils/clean')
 
-describe('HTTP API', () => {
+describe('HTTP GATEWAY', () => {
   const repoExample = path.join(__dirname, '../go-ipfs-repo')
   const repoTests = path.join(__dirname, '../repo-tests-run')
 
@@ -36,20 +36,20 @@ describe('HTTP API', () => {
     })
   })
 
-  describe('## http-api spec tests', () => {
+  describe('## http-gateway spec tests', () => {
     fs.readdirSync(path.join(__dirname, '/spec'))
       .forEach((file) => require('./spec/' + file)(http))
   })
 
-  describe('## interface tests', () => {
-    fs.readdirSync(path.join(__dirname, '/interface'))
-      .forEach((file) => require('./interface/' + file))
-  })
-
-  describe('## custom ipfs-api tests', () => {
-    const ctl = APIctl('/ip4/127.0.0.1/tcp/6001')
-
-    fs.readdirSync(path.join(__dirname, '/over-ipfs-api'))
-      .forEach((file) => require('./over-ipfs-api/' + file)(ctl))
-  })
+  // describe('## interface tests', () => {
+  //   fs.readdirSync(path.join(__dirname, '/interface'))
+  //     .forEach((file) => require('./interface/' + file))
+  // })
+  //
+  // describe('## custom ipfs-api tests', () => {
+  //   const ctl = APIctl('/ip4/127.0.0.1/tcp/6001')
+  //
+  //   fs.readdirSync(path.join(__dirname, '/over-ipfs-api'))
+  //     .forEach((file) => require('./over-ipfs-api/' + file)(ctl))
+  // })
 })
