+++ test/utils/ipfs-factory-daemon/index.js
@@ -3,7 +3,7 @@
 const PeerId = require('peer-id')
 const IPFSAPI = require('ipfs-api')
 const clean = require('../clean')
-const HttpApi = require('../../../src/http-api')
+const HttpApi = require('../../../src/http')
 const series = require('async/series')
 const eachSeries = require('async/eachSeries')
 const defaultConfig = require('./default-config.json')
